LayoutTests/inspector/console-log-syntax-error.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="../http/tests/inspector/console-tests.js"></script>
<script src="resources/syntax-error.js"></script>
<script>
function doit()
{
    dumpConsoleMessages();
}
</script>
</head>
<body onload="onload()">
<p>
Tests that syntax errors are logged into console and doesn't cause browser crash.
</p>
</body>
</html>
